Split the promo-box tests out of the catch-all form test file
The form suite grew past the 400-line limit with the promo pre-fill tests.
They now have their own file under the form mirror path, so the mutation
gate still owns them for form.tsx.
